3. Web Api/requirements.txt
flask
flask-restful
flask-sqlalchemy
pymysql
flask-migrate
flask-script
flask-jwt-extended
password_strength
requests
pytest
flask_cors

